Copied sleep function into one circuit
Copied and pasted code from Ben Duggan's ETAG Revision repository to enable sleep function. Polling one antenna for 3000ms. When in sleep function, read tags will not appear on serial monitor but will be logged in the data file. While in sleep mode, board is supposed to be using 864 microA.