Reverted to antenna sleep, no noise blips
Board will be awake but antennas will be asleep, will need to change poll time based on how long battery will stay awake (for Kinkoo battery, it will stay on for 9 s if the power button is pressed and it is not plugged into anything). Had no noise blips when board and mp3 player were connected to Kinkoo battery.