Converted time variables to min, added wakeUpTime variable
When time variables (startTime and endTime) were in seconds the reader would sometimes skip them since they were so precise. Converted the time variables back to minutes, although the sound will not start exactly on 5:30:00 but a few seconds after (e.c. 5:30:10).

Added a wakeUpTime variable so when the serial mp3 player wakes up the noise is not choppy. Having the two separate startTime and wakeUpTime variables leads to smooth, uninterrupted noise at 5:30